backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
scikit-learn==1.3.2
tensorflow==2.15.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.2
pyjwt==2.8.0
requests==2.31.0
